requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
pandas==2.2.3
numpy==2.2.3
httpx==0.28.1          # async HTTP client (replaces requests)
geopy==2.4.1
scipy==1.15.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
uvicorn[standard]==0.34.0  # ASGI server for async support
gunicorn==23.0.0
whitenoise==6.9.0
drf-spectacular==0.28.0   # OpenAPI 3.0 schema + Swagger UI + ReDoc
